# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game Library API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/library": {
            "get": {
                "tags": [
                    "Library"
                ],
                "summary": "Pobierz listę gier z biblioteki (API)",
                "operationId": "d469bcec0ae40b110c3dc583a00a2e6a",
                "responses": {
                    "200": {
                        "description": "Lista gier w formacie JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "games": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "rawg_game_id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "nullable": true
                                                    },
                                                    "cover_url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Library"
                ],
                "summary": "Dodaj grę do biblioteki (API)",
                "operationId": "baaa832b661f395a3380a5b8d1413b49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rawg_game_id",
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "rawg_game_id": {
                                        "type": "integer",
                                        "example": 3498
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "The Witcher 3: Wild Hunt"
                                    },
                                    "cover_url": {
                                        "type": "string",
                                        "example": "https://image.url",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "to_play",
                                            "playing",
                                            "finished"
                                        ],
                                        "example": "playing"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 10,
                                        "minimum": 1,
                                        "example": 9,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gra dodana"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/library/{id}": {
            "put": {
                "tags": [
                    "Library"
                ],
                "summary": "Zaktualizuj grę w bibliotece (API)",
                "operationId": "dc5b2071ced380e087ee7316acf9fc35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "to_play",
                                            "playing",
                                            "finished"
                                        ]
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 10,
                                        "minimum": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gra zaktualizowana"
                    },
                    "403": {
                        "description": "Nie twoja gra"
                    },
                    "404": {
                        "description": "Gra nie znaleziona"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Library"
                ],
                "summary": "Usuń grę z biblioteki (API)",
                "operationId": "f6a1a9a313d231db2ef64aa8b779628f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Gra usunięta"
                    },
                    "403": {
                        "description": "Nie twoja gra"
                    },
                    "404": {
                        "description": "Gra nie znaleziona"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}